third_party/WebKit/LayoutTests/fast/events/pointerevents/mouse-pointer-event-properties.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/input-modifiers.js"></script>
<style>
div.box {
  margin: 10px;
  padding: 50px;
  float: left;
}
</style>

<div id="target" class="box" style="background-color:red">
</div>

<div id="console"></div>

<script>
window.name = "mainWindow";
description("Verifies that pointer event parameters are correct when fired through mouse events.");

var checkKeyModifiers = false;

var pointerType = "";
var penId = 0;
var penPressure = 0;
var penTiltX = 0;
var penTiltY = 0;

var testEventList = ["mouseenter", "mouseleave", "mouseover", "mouseout", "mouseup", "mousedown", "mousemove",
                   "pointerenter", "pointerleave", "pointerover", "pointerout", "pointerup", "pointerdown", "pointermove"];
var lastPointerEvents = [];

var numericAttributes = [
    "clientX",
    "clientY",
    "layerX",
    "layerY",
    "movementX",
    "movementY",
    "offsetX",
    "offsetY",
    "pageX",
    "pageY",
    "screenX",
    "screenY",
    "x",
    "y",
    "button",
    "buttons",
    "pressure",
    "tiltX",
    "tiltY",
    "width",
    "height",
];

function init() {
  var targetDiv = document.getElementById("target");

  testEventList.forEach(function(eventName) {

    targetDiv.addEventListener(eventName, function(event) {

      debug("Received " + event.type);

      if (event.type.startsWith("pointer"))
        lastPointerEvents.push(event);
      else {
        if (event.type == "mouseout" || event.type == "mouseover") {
          shouldBeEqualToNumber("lastPointerEvents.length", 2);
        } else {
          shouldBeEqualToNumber("lastPointerEvents.length", 1);
        }
        shouldBeEqualToString("lastPointerEvents[0].type", event.type.replace("mouse", "pointer"));

        if (!checkKeyModifiers) {

          if (lastPointerEvents[0].type=="pointerenter" || lastPointerEvents[0].type=="pointerleave") {
            shouldBeFalse("lastPointerEvents[0].bubbles");
            shouldBeFalse("lastPointerEvents[0].cancelable");
          } else {
            shouldBeTrue("lastPointerEvents[0].bubbles");
            shouldBeTrue("lastPointerEvents[0].cancelable");
          }

          shouldBeEqualToNumber("lastPointerEvents[0].pointerId",
              (pointerType == "mouse")? 1 : penId);
          shouldBeEqualToString("lastPointerEvents[0].pointerType", pointerType);
          shouldBeTrue("lastPointerEvents[0].isPrimary");

          numericAttributes.forEach(function(attribute) {
            var expectedValue = eval("event." + attribute);
            if (attribute == "button" && event.type != "mousedown" && event.type != "mouseup")
              expectedValue = -1;
            else if (attribute == "width" || attribute == "height")
              expectedValue = 1;
            else if (attribute == "pressure") {
              if (pointerType == "mouse")
                expectedValue = (event.buttons == 0)? 0.0 : 0.5;
              else
                expectedValue = penPressure;
            }
            else if (attribute == "tiltX")
              expectedValue = (pointerType == "mouse")? 0 : penTiltX;
            else if (attribute == "tiltY")
              expectedValue = (pointerType == "mouse")? 0 : penTiltY;

            shouldBeEqualToNumber("lastPointerEvents[0]." + attribute, expectedValue);
          });
          shouldBeEqualToString("lastPointerEvents[0].view.name", "mainWindow");

        } else {

          forEachModifier(function(attr, modifierName) {
            var getModifierStateStr = ".getModifierState('" + modifierName + "');"
            if (eval("event" + getModifierStateStr))
              shouldBeTrue("lastPointerEvents[0]" + getModifierStateStr);
            else
              shouldBeFalse("lastPointerEvents[0]" + getModifierStateStr);
          });

        }

        lastPointerEvents.splice(0, 1);
      }
    });

  });
}

function runMouseTests(x, y) {
  debug("===== mouse tests =====");
  pointerType = "mouse";

  debug("--- move mouse into target ---");
  eventSender.mouseMoveTo(x + 5, y + 5);
  debug("");

  debug("--- move within target ---");
  eventSender.mouseMoveTo(x + 7, y + 15);
  eventSender.mouseMoveTo(x + 5, y + 5);
  debug("");

  debug("--- click each button ---");
  for (var button = 0; button <=2; button++) {
    eventSender.mouseDown(button);
    eventSender.mouseUp(button);
    // TODO(crbug.com/548226): Investigate missing events in win_chromium_rel_ng on 3rd down/up.
  }
  debug("");

  debug("--- click with each modifier ---");
  checkKeyModifiers = true;
  forEachModifier(function(attr, modifierName, eventSenderName) {
    eventSender.mouseDown(0, [eventSenderName]);
    eventSender.mouseUp(0, [eventSenderName]);
  });
  checkKeyModifiers = false;
  debug("");

  debug("--- move mouse out of target ---");
  eventSender.mouseMoveTo(x - 5, y - 5);

  debug("");
}

function runPenTests(x, y) {
  debug("===== pen tests =====");
  pointerType = "pen";
  penId = 2;
  penPressure = 0.0;
  penTiltX = 0;
  penTiltY = 0;

  debug("--- move pen into target ---");
  eventSender.mouseMoveTo(x + 5, y + 5, [], "pen", penId, penPressure, penTiltX, penTiltY);
  debug("");

  debug("--- move within target & tap ---");
  penTiltX = 45;
  penTiltY = -34;
  eventSender.mouseMoveTo(x + 15, y + 15, [], "pen", penId, penPressure, penTiltX, penTiltY);
  penPressure = 0.75;
  eventSender.mouseDown(0, [], "pen", penId, penPressure, penTiltX, penTiltY);
  penPressure = 0.0;
  eventSender.mouseUp(0, [], "pen", penId, penPressure, penTiltX, penTiltY);

  debug("--- move pen out of target ---");
  eventSender.mouseMoveTo(x - 5, y - 5, [], "pen", penId, penPressure, penTiltX, penTiltY);

  debug("");
}

function runAllTests() {
  var rect = document.getElementById("target").getBoundingClientRect();

  runMouseTests(rect.left, rect.top);
  runPenTests(rect.left, rect.top);
}

init();
if (window.eventSender)
  runAllTests();
else
  debug("This test requires eventSender");

</script>
